# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management Service",
        "description": "Defines API endpoints for managing translations with features like creation, updates, search, and JSON export. Focuses on scalability, token-based authentication.",
        "contact": {
            "email": "mhaideraziz22@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://tms-app.test"
        }
    ],
    "paths": {
...
        "/api/translations": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create Translation",
                "description": "Creates a new translation with a unique key, optional tags, and multiple language entries.",
                "operationId": "fe38d8cf02ff5c92d2cd8d5dfb095e5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "languages"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "greeting"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "common",
                                        "nullable": true
                                    },
                                    "languages": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "locale": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Hello"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Translation created successfully"
                                        },
                                        "translation": {
                                            "properties": {
                                                "key": {
                                                    "type": "string",
                                                    "example": "greeting"
                                                },
                                                "tags": {
                                                    "type": "string",
                                                    "example": "common",
                                                    "nullable": true
                                                },
                                                "languages": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "locale": {
                                                                "type": "string",
                                                                "example": "en"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "Hello"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "key": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The key field is required."
                                                    }
                                                },
                                                "languages": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The languages field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/translations/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search Translations",
                "description": "Search translations by key, tags, or content. Filters are optional.",
                "operationId": "37125c7da8b2fa8e457148f15ccd2698",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Translation key to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "greeting"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tags associated with the translation",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "common"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content to search within the translations' languages",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Hello"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "greeting"
                                            },
                                            "tags": {
                                                "type": "string",
                                                "example": "common",
                                                "nullable": true
                                            },
                                            "languages": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "locale": {
                                                            "type": "string",
                                                            "example": "en"
                                                        },
                                                        "content": {
                                                            "type": "string",
                                                            "example": "Hello"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "key": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The key field must be a string."
                                                    }
                                                },
                                                "tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The tags field must be a string."
                                                    }
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The content field must be a string."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer token for authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API key for accessing the API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}